Removes `log` and `env_logger` (and small fixes) (#12)
Removes log and env_logger because we are not a daemon or service, so that style of log output is not helpful. Instead I added the same 'level style silencing' to our cli_* printing macros. This means all of our cli_* print macros work with the --verbose and --quiet unlike the previous log/env_logger style macros.