implement `formation status` command (#115)
This PR implements the `formation status` command :tada:

It checks the number of running containers if all containers are above the minimum set, it lists the Formation as UP in a pretty way. If some containers have too few, or there are containers reporting exit with an error it will show the configuration as `Degraded`. The containers are down, it will show as `DOWN`.

A Formation shows the same status as the lowest status of it's combined configurations. So a Formation with multiple configurations will show UP only if all configurations show UP. If one shows `Degraded` or `Down` it the Formation as a whole will show the lowest status.

Here's what it looks like with two formations:

<img width="586" alt="Screen Shot 2022-04-15 at 9 00 40 AM" src="https://user-images.githubusercontent.com/6942134/163574712-4398e0ad-8027-46f5-98b2-68bacc9d845e.png">