Use UUIDv4 for now (#297)
UUIDv7 is currently unstable in the IETF. This means there are various
differing implementations. The Rust `uuid` crate follows the RFC draft
as written, and is up to date with that draft, however they require both
the use of a compile time feature `v7` and a `rustc` flag (`--cfg
uuid_unstable`) in order to use UUIDv7. While doable, this has build
system implications.

Also, the property of being timestamp sortable is not a hard requirment
for our system (yet?). If that becomes a hard requirement, and the IETF
RFC stablises (supposedly in Mar 2023?) we can re-address adding UUIDv7.
For now, we are using random (v4) UUIDs.